fix: delete personal info component, integrate in up detail
