+++ src/app/user-profile/user-profile-detail/user-profile-detail.component.scss
